converter: box a named untyped-int constant to nint even in a variadic ...any slot
fmt.Sprintf("%d", fsize+1), where fsize is a package/function const declared without
an explicit type (const fsize = 5), rendered fsize as a golib UntypedInt-typed C#
variable. The variadic-...any box-cast skip (added to avoid redundant (nint) casts on
plain int literals passed to fmt/print/log calls) also swallowed this shape, so the
UntypedInt STRUCT boxed directly to object instead of a CLR integer. fmt's printArg
type-switch (print.cs) doesn't recognize UntypedInt, falls back to reflection, and
prints it as a two-field struct instead of the plain value -- go/token's
TestIssue57490 saw "{6 %!d(bool=false)}" instead of "6" against the full-conversion
fmt (go-src-converted/fmt), the root cause blocking that test (30/31 now; the last
failure, TestSerialization, needs the gob/reflect increment 2, out of scope here).

exprInvolvesUntypedIntConst (convCallExpr.go) walks the argument's AST for an Ident
resolving to a *types.Const whose OWN declared type stays UntypedInt/UntypedRune --
not info.Types[arg]'s post-default type, which reports plain `int` for both a literal
and a named const alike -- and re-enables the nint cast for exactly that argument,
leaving the literal-only fast path (fmt.Println(42)) untouched.

Verified with a standalone harness (documented in the reference doc) that references
go-src-converted/fmt + core/golib directly (bypassing the baseline core/fmt stub,
which happens to mask this bug via UntypedInt.ToString()): fmt.Sprintf("%d", fsize+1)
now prints "6", matching `go run`, both before and after the fix confirmed via a
pre/post reconvert diff.

Isolation-reconvert-diff across the full 302-package corpus (pre-fix vs post-fix,
independent of the tree's pre-existing staleness) surfaced exactly 10 affected files,
all the same latent %! bad-verb bug in real stdlib error/panic messages (go/types,
crypto/tls, internal/coverage, internal/trace, math/big); all 10 build clean and are
re-baselined here alongside the fix. Two behavioral goldens (IotaEnum, UntypedConstDefine)
also re-baseline for the same reason; UntypedConstDefine gains a guard line pinning this
exact shape (fmt.Sprintf("%d", fsize+1) via a package-level const), Output-compared vs
`go run`.

Gates: check-no-regression.ps1 byte-identical except the two inspected/intended
re-baselines; full behavioral suite 474/474 across all four phases; the 10
corpus-diff packages build clean; go/token pipeline re-run confirms 30/31 (up from
29/31), TestIssue57490 passing.

A related, separate divergence surfaced by the same harness -- a bare int LITERAL
still boxes as System.Int32, and the full-conversion fmt's badVerb/%T path names it
via reflect.TypeOf(arg).String() (go-src-converted/reflect + golib GoReflect), which
reports "int32" instead of Go's "int" (GetGoTypeName, used by TryTypeAssert panics and
the baseline stub fmt, already normalizes this correctly) -- is documented in the
reference doc as a known, disclosed, NOT-fixed-here gap; closing it would need a
GoReflect-side fix, not a wider variadic-cast footprint.

Docs: ConversionStrategies-Reference.md's Empty Interface (any) section and
ConversionStrategies.md's summary both updated to describe the refined variadic
carve-out and the residual int32-vs-int gap.

(cherry picked from commit 9129d40a15365ec6618fa89f5c730a8b0cde8cce)

+++ src/Tests/Behavioral/IotaEnum/IotaEnum.cs
@@ -220,7 +220,7 @@ internal static void Main() {
     fmt.Printf("Slice Kind Value: %s [%d]\n"u8, Slice.String(), (nint)(uint8)Slice);
     fmt.Println(seqFirst, seqSecond, seqAgain);
     fmt.Println(wideFirst, wideNext);
-    fmt.Println(rawZero, rawOne);
+    fmt.Println((nint)(rawZero), (nint)(rawOne));
     fmt.Println(stateMachine(0));
     fmt.Println(stateMachine(1));
     fmt.Println(stateMachine(2));
